Add script chaining raw files in readout.exe
As readout.exe can handle with one equipment only one
raw file at the time, the idea to run over many files is to
have a script generator creating configs with the raw file
as parameter and starting automatically readout.exe
once the previous process runs into timeout. The
consumer application should not be affected as it
listens on the same channel.